backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
supabase==2.0.3
python-dotenv==1.0.0
httpx==0.25.1
mistralai==0.0.11
python-dateutil==2.8.2
APScheduler==3.10.4
loguru==0.7.2 